Revert "Test without replacing the manifest file"
This reverts commit a98512ecd34be62f6696da3b5f6b17f33a1299be.

+++ ci/init_build.sh
@@ -80,21 +80,21 @@ yarn kbn bootstrap --allow-root
 echo -e "\e[0Ksection_end:`date +%s`:yarn_bootstrap\r\e[0K"
 
 
-#start_section replace_kbn_ui_shared_deps_npm_manifest_json "Replace bazel-bin/../kbn-ui-shared-deps-npm-manifest.json with file from kibana release"
-#
-#echo "Will download $SF_RELEASE_PACKAGE_URL"
-#
-#curl --fail $SF_RELEASE_PACKAGE_URL -o kibana-$SF_VERSION.tar.gz
-#tar -xf kibana-$SF_VERSION.tar.gz --transform 's/.*\///g' --wildcards --no-anchored 'kbn-ui-shared-deps-npm-manifest.json'
-#
-#if [ ! -f bazel-bin/src/platform/packages/private/kbn-ui-shared-deps-npm/shared_built_assets/kbn-ui-shared-deps-npm-manifest.json ]; then
-#  echo "File bazel-bin/src/platform/packages/kbn-ui-shared-deps-npm/shared_built_assets/kbn-ui-shared-deps-npm-manifest.json not found"
-#  exit 1
-#fi
-#mv kbn-ui-shared-deps-npm-manifest.json bazel-bin/src/platform/packages/private/kbn-ui-shared-deps-npm/shared_built_assets/kbn-ui-shared-deps-npm-manifest.json
-#rm -rf kibana-$SF_VERSION.tar.gz
-#
-#end_section replace_kbn_ui_shared_deps_npm_manifest_json
+start_section replace_kbn_ui_shared_deps_npm_manifest_json "Replace bazel-bin/../kbn-ui-shared-deps-npm-manifest.json with file from kibana release"
+
+echo "Will download $SF_RELEASE_PACKAGE_URL"
+
+curl --fail $SF_RELEASE_PACKAGE_URL -o kibana-$SF_VERSION.tar.gz
+tar -xf kibana-$SF_VERSION.tar.gz --transform 's/.*\///g' --wildcards --no-anchored 'kbn-ui-shared-deps-npm-manifest.json'
+
+if [ ! -f bazel-bin/src/platform/packages/private/kbn-ui-shared-deps-npm/shared_built_assets/kbn-ui-shared-deps-npm-manifest.json ]; then
+  echo "File bazel-bin/src/platform/packages/kbn-ui-shared-deps-npm/shared_built_assets/kbn-ui-shared-deps-npm-manifest.json not found"
+  exit 1
+fi
+mv kbn-ui-shared-deps-npm-manifest.json bazel-bin/src/platform/packages/private/kbn-ui-shared-deps-npm/shared_built_assets/kbn-ui-shared-deps-npm-manifest.json
+rm -rf kibana-$SF_VERSION.tar.gz
+
+end_section replace_kbn_ui_shared_deps_npm_manifest_json
 
 mkdir -p plugins/search-guard
 cp -a "../babel.config.js" plugins/search-guard
